- Introduce singleton.py - Refactor device_registry.py
