+++ src/utils/__init__.py
@@ -1,0 +1,3 @@
+from .singleton import Singleton
+
+__all__ = [Singleton]
